divide accountDal to submodules
