+++ server/api/account/dal/createUser.js
@@ -3,7 +3,7 @@ const config = require('georap-config');
 const db = require('georap-db');
 const bcrypt = require('bcryptjs');
 
-exports.createUser = function (username, email, password, callback) {
+module.exports = function (username, email, password, callback) {
   // Create non-admin active user. Does not check if user exists.
   //
   // Parameters:
@@ -53,38 +53,5 @@ exports.createUser = function (username, email, password, callback) {
       // User inserted successfully
       return callback();
     });
-  });
-};
-
-exports.markLogin = function (username, callback) {
-  // Set last login time to the current time.
-  // If user is not found, fails silently.
-  //
-  // Parameters:
-  //   username
-  //     string
-  //   callback
-  //     function (err)
-  //
-  const users = db.collection('users');
-
-  const q = { name: username };
-  const u = {
-    $set: {
-      loginAt: (new Date()).toISOString(),
-    },
-  };
-
-  users.updateOne(q, u, (err, result) => {
-    if (err) {
-      return callback(err);
-    }
-
-    if (result.matchedCount === 0) {
-      // No users found
-      return callback();
-    }
-
-    return callback();
   });
 };
