+++ server/api/account/dal/index.js
@@ -1,0 +1,2 @@
+exports.createUser = require('./createUser');
+exports.markLogin = require('./markLogin');
